requirements.txt
flask
flask-cors
opencv-python
pillow
numpy
python-dotenv
langchain-google-genai 
google-generativeai